requirements.txt
Django==2.0.7
djangorestframework==3.8.2
celery==4.2.0
-e lcs
